Refactoring the Game page a bit; moved ai analysis prediction chart into the Game section
